+++ src/lib.rs
@@ -39,6 +39,7 @@ mod connection;
 mod device;
 mod error;
 mod flow;
+pub mod is05;
 mod media;
 mod node;
 mod receiver;